removed admin login
removed the section of code which was stopping me from using the same login, thought it would be nice to have to logins when in reality not needed. This was the issue encounted before which flustered me however was a simple mistake. Additionally i have also updating the user_info route to allow new signups to see there account info eventhough they havnt inputted there data.